Migrate pages examples to MDX
- No more storiesOf
- Moved pages to _internal/pages
- Added some noHorizontalSpacing to fix some layout issues
- Added baseUrl to tsconfig to avoid LONG relative imports
