+++ src/_internals/pages/login/Email.story.mdx
@@ -1,0 +1,41 @@
+import { Meta, Preview, Story } from '@storybook/addon-docs/blocks'
+
+import { Button } from 'button'
+import { SpacingDivider, SpacingDividerSize } from 'divider/spacingDivider'
+import { ItemAction } from 'itemAction'
+import { ItemCheckbox } from 'itemCheckbox'
+import { BottomContent, Content, MainContent } from 'layout/content'
+import { BaseSection } from 'layout/section/baseSection'
+import { TextField } from 'textField'
+import { TheVoice } from 'theVoice'
+
+<Meta title="Pages|Login" />
+
+<Preview>
+  <Story name="Login page">
+    <React.Fragment>
+      <MainContent>
+        <Content>
+          <BaseSection noHorizontalSpacing>
+            <TheVoice>What&apos;s your email and password?</TheVoice>
+          </BaseSection>
+          <BaseSection>
+            <TextField />
+            <SpacingDivider size={SpacingDividerSize.MEDIUM} />
+            <TextField />
+            <SpacingDivider size={SpacingDividerSize.SMALL} />
+          </BaseSection>
+          <BaseSection noHorizontalSpacing>
+            <ItemCheckbox name="rememberme" labelTitle="Remember me" />
+            <ItemAction href="#" action="Forgot password?" />
+          </BaseSection>
+        </Content>
+        <BottomContent>
+          <div style={{ textAlign: 'center' }}>
+            <Button>Log in</Button>
+          </div>
+        </BottomContent>
+      </MainContent>
+    </React.Fragment>
+  </Story>
+</Preview>
