+++ src/_internals/pages/search/Results.story.mdx
@@ -1,0 +1,11 @@
+import { Meta, Preview, Story } from '@storybook/addon-docs/blocks'
+
+import { Tabs } from './Tabs'
+
+<Meta title="Pages|Search results" />
+
+<Preview>
+  <Story name="With tabs">
+    <Tabs />
+  </Story>
+</Preview>
