+++ src/_internals/pages/yourrides/Rides.story.mdx
@@ -1,0 +1,14 @@
+import { Meta, Preview, Story } from '@storybook/addon-docs/blocks'
+
+import { Rides, RidesHistory } from './Rides'
+
+<Meta title="Pages|Your rides" />
+
+<Preview>
+  <Story name="Active rides">
+    <Rides />
+  </Story>
+  <Story name="Inactive rides">
+    <RidesHistory />
+  </Story>
+</Preview>
