+++ tsconfig.json
@@ -15,7 +15,8 @@
     "allowJs": false,
     "declaration": true,
     "noEmitOnError": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": "./src"
   },
   "include": ["node_modules/@types", "src/**/*.tsx", "src/**/*.ts"],
   "exclude": ["build", "node_modules", "src/**/*.unit.tsx", "tools"]
